Seperated the testing of predefined distributions.
